Address comment

+++ pkg/cluster/manager.go
@@ -47,6 +47,10 @@ import (
 var (
 	errNSDeploy            = errorx.NewNamespace("deploy")
 	errDeployNameDuplicate = errNSDeploy.NewType("name_dup", errutil.ErrTraitPreCheck)
+
+	errNSRename              = errorx.NewNamespace("rename")
+	errorRenameNameNotExist  = errNSRename.NewType("name_not_exist", errutil.ErrTraitPreCheck)
+	errorRenameNameDuplicate = errNSRename.NewType("name_dup", errutil.ErrTraitPreCheck)
 )
 
 // Manager to deploy a cluster.
@@ -524,6 +528,17 @@ func (m *Manager) EditConfig(clusterName string, skipConfirm bool) error {
 
 // Rename the cluster
 func (m *Manager) Rename(clusterName string, opt operator.Options, newName string) error {
+	if !utils.IsExist(m.specManager.Path(clusterName)) {
+		return errorRenameNameNotExist.
+			New("Cluster name '%s' not exist", clusterName).
+			WithProperty(cliutil.SuggestionFromFormat("Please double check your cluster name"))
+	}
+	if utils.IsExist(m.specManager.Path(newName)) {
+		return errorRenameNameDuplicate.
+			New("Cluster name '%s' is duplicated", newName).
+			WithProperty(cliutil.SuggestionFromFormat("Please specify another cluster name"))
+	}
+
 	if err := os.Rename(m.specManager.Path(clusterName), m.specManager.Path(newName)); err != nil {
 		return perrs.AddStack(err)
 	}
